requirements.txt
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.7
django-admin-thumbnails==0.2.6
django-allauth==0.52.0
django-ckeditor==6.5.1
django-colorfield==0.8.0
django-import-export==3.2.0
django-js-asset==2.0.0
django-modeltranslation==0.18.9
django-mptt==0.14.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
MarkupPy==1.14
multidict==6.0.4
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.4.0
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
Pyrogram==2.0.102
PySocks==1.7.1
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tablib==3.4.0
Telethon==1.27.0
typing_extensions==4.5.0
urllib3==1.26.14
xlrd==2.0.1
xlwt==1.3.0
yarl==1.8.2
